[generator] Refactoring: some code was put to features_processing_helpers.hpp; renamed method
