+++ generator/features_processing_helpers.hpp
@@ -1,0 +1,28 @@
+#pragma once
+
+#include "generator/feature_builder.hpp"
+
+#include "base/thread_safe_queue.hpp"
+
+#include <cstddef>
+#include <utility>
+#include <string>
+#include <vector>
+
+namespace generator
+{
+size_t static const kAffilationsBufferSize = 512;
+
+struct ProcessedData
+{
+  explicit ProcessedData(feature::FeatureBuilder::Buffer && buffer,
+                         std::vector<std::string> && affiliations)
+    : m_buffer(std::move(buffer)), m_affiliations(std::move(affiliations)) {}
+
+  feature::FeatureBuilder::Buffer m_buffer;
+  std::vector<std::string> m_affiliations;
+};
+
+using FeatureProcessorChank = base::threads::DataWrapper<std::vector<ProcessedData>>;
+using FeatureProcessorQueue = base::threads::ThreadSafeQueue<FeatureProcessorChank>;
+}  // namespace generator
