+++ generator/processor_coastline.hpp
@@ -18,7 +18,7 @@ public:
   std::shared_ptr<FeatureProcessorInterface> Clone() const override;
 
   void Process(feature::FeatureBuilder & feature) override;
-  void Flush() override;
+  void Finish() override;
 
   void Merge(FeatureProcessorInterface const & other) override;
   void MergeInto(ProcessorCoastline & other) const override;
