+++ generator/processor_country.cpp
@@ -37,7 +37,7 @@ void ProcessorCountry::Process(feature::FeatureBuilder & feature)
   m_affilationsLayer->AddBufferToQueueIfFull(m_queue);
 }
 
-void ProcessorCountry::Flush()
+void ProcessorCountry::Finish()
 {
   m_affilationsLayer->AddBufferToQueue(m_queue);
 }
