+++ generator/processor_country.hpp
@@ -27,7 +27,7 @@ public:
   std::shared_ptr<FeatureProcessorInterface> Clone() const override;
 
   void Process(feature::FeatureBuilder & feature) override;
-  void Flush() override;
+  void Finish() override;
 
   void Merge(FeatureProcessorInterface const & other) override;
   void MergeInto(ProcessorCountry & other) const override;
