+++ generator/processor_interface.hpp
@@ -29,7 +29,7 @@ public:
 
   // This method is used by OsmTranslator to pass |fb| to Processor for further processing.
   virtual void Process(feature::FeatureBuilder & fb) = 0;
-  virtual void Flush() = 0;
+  virtual void Finish() = 0;
 
   virtual void Merge(FeatureProcessorInterface const &) = 0;
 
@@ -42,21 +42,4 @@ public:
 private:
   void FailIfMethodUnsupported() const { CHECK(false, ("This method is unsupported.")); }
 };
-
-size_t static const kAffilationsBufferSize = 128;
-
-struct ProcessedData
-{
-  explicit ProcessedData(feature::FeatureBuilder::Buffer && buffer, std::vector<std::string>  && affiliations)
-    : m_buffer(std::move(buffer)), m_affiliations(std::move(affiliations))
-  {
-  }
-
-  feature::FeatureBuilder::Buffer m_buffer;
-  std::vector<std::string> m_affiliations;
-};
-
-using FeatureProcessorChank = base::threads::DataWrapper<std::vector<ProcessedData>>;
-using FeatureProcessorQueue = base::threads::ThreadSafeQueue<FeatureProcessorChank>;
-
 }  // namespace generator
