+++ generator/processor_simple.hpp
@@ -21,7 +21,7 @@ public:
   std::shared_ptr<FeatureProcessorInterface> Clone() const override;
 
   void Process(feature::FeatureBuilder & fb) override;
-  void Flush() override;
+  void Finish() override;
 
   void Merge(FeatureProcessorInterface const & other) override;
   void MergeInto(ProcessorSimple & other) const override;
